Fix race condition causing use-after-free bugs
When the plugin is instantiated, it starts loading preset data on a
separate thread. If the plugin is destroyed before the thread finishes,
the thread ends up accessing memory that was already deallocated,
leading to use-after-free memory violations.

This issue is particularly likely to occur in `juce_vst3_helper`, which
instantiates the plugin in order to call `getVST3ClientExtensions()` but
then almost immediately destroys it. The preset loading thread is likely
still running at that point.

This commit modifies the destructor of `VASTPresetData` so that it
waits for the thread to finish, if necessary.